feat(builder): add source-build builer plugin
