--- .changeset/tame-camels-switch.md
@@ -1,6 +1,0 @@
----
-'@modern-js/app-tools': minor
----
-
-feat(app-tools): Add sourceBuild default configuration, register builder-plugin-source-build builder plugin
-feat(app-tools): 新增 sourceBuild 默认配置, 注册 builder-plugin-source-build builder 插件
