+++ .changeset/weak-experts-perform.md
@@ -1,6 +1,6 @@
 ---
-'@modern-js/builder-plugin-source-build': minor
+'@modern-js/builder': minor
 ---
 
-feat(builder-plugin-source-build): New builder plug-in to support source code building
-feat(builder-plugin-source-build): 新增支持源码构建的 builder 插件
+feat(builder): New builder plug-in to support source code building
+feat(builder): 新增支持源码构建的 builder 插件
