+++ .changeset/yellow-pumas-cheer.md
@@ -1,0 +1,6 @@
+---
+'@modern-js/builder-shared': patch
+---
+
+feat(builder-shared): add sourceBuild type definition
+feat(builder-shared): 新增 sourceBuild 类型定义
