+++ packages/builder/builder-webpack-provider/src/shared/plugin.ts
@@ -68,6 +68,7 @@ export const applyDefaultPlugins = (plugins: Plugins) =>
     plugins.inlineChunk(),
     plugins.assetsRetry(),
     plugins.externals(),
+    plugins.sourceBuild(),
     import('../plugins/performance').then(m => m.builderPluginPerformance()),
     import('../plugins/lazyCompilation').then(m =>
       m.builderPluginLazyCompilation(),
