+++ packages/builder/builder/src/index.ts
@@ -1,6 +1,8 @@
 export { createBuilder } from './createBuilder';
 export { mergeBuilderConfig } from '@modern-js/builder-shared';
 
+export { builderPluginSourceBuild } from './plugins/sourceBuild';
+
 export type {
   BuilderMode,
   BuilderEntry,
