--- packages/builder/plugin-source-build/README.md
@@ -1,27 +1,0 @@
-
-<p align="center">
-  <a href="https://modernjs.dev" target="blank"><img src="https://lf3-static.bytednsdoc.com/obj/eden-cn/ylaelkeh7nuhfnuhf/modernjs-cover.png" width="300" alt="Modern.js Logo" /></a>
-</p>
-<p align="center">
-  Modern.js Builder
-  <br/>
-  <a href="https://modernjs.dev/builder/en" target="blank">
-    modernjs.dev/builder
-  </a>
-</p>
-<p align="center">
-  A build engine for web development.
-</p>
-
-# @modern-js/builder-plugin-source-build
-
-This package is the source build plugin of Modern.js Builder.
-
-## Getting Started
-
-- [Introduction](https://modernjs.dev/builder/en/guide/introduction.html)
-- [API References](https://modernjs.dev/builder/en/api/)
-
-## Contributing
-
-- [Contributing Guide](https://github.com/web-infra-dev/modern.js/blob/main/CONTRIBUTING.md)
