--- packages/builder/plugin-source-build/package.json
@@ -1,58 +1,0 @@
-{
-  "name": "@modern-js/builder-plugin-source-build",
-  "description": "Build source Code plugin for builder in Modern.js",
-  "homepage": "https://modernjs.dev",
-  "bugs": "https://github.com/web-infra-dev/modern.js/issues",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/web-infra-dev/modern.js",
-    "directory": "packages/builder/plugin-node-polyfill"
-  },
-  "license": "MIT",
-  "keywords": [
-    "react",
-    "framework",
-    "modern",
-    "modern.js"
-  ],
-  "engines": {
-    "node": ">=14.0.0"
-  },
-  "version": "2.22.0",
-  "jsnext:source": "./src/index.ts",
-  "types": "./src/index.ts",
-  "main": "./dist/index.js",
-  "module": "./dist/index.js",
-  "exports": {
-    ".": {
-      "jsnext:source": "./src/index.ts",
-      "default": "./dist/index.js"
-    }
-  },
-  "scripts": {
-    "prepublishOnly": "only-allow-pnpm",
-    "new": "modern new",
-    "build": "tsc",
-    "dev": "tsc --watch",
-    "test": "vitest run",
-    "test:watch": "vitest dev --no-coverage"
-  },
-  "dependencies": {
-    "@modern-js/builder-shared": "workspace:*",
-    "@modern-js/monorepo-utils": "workspace:*"
-  },
-  "devDependencies": {
-    "@modern-js/builder": "workspace:*",
-    "@modern-js/builder-rspack-provider": "workspace:*",
-    "@modern-js/builder-webpack-provider": "workspace:*",
-    "@modern-js/utils": "workspace:*",
-    "@scripts/vitest-config": "workspace:*",
-    "typescript": "^5"
-  },
-  "sideEffects": false,
-  "publishConfig": {
-    "registry": "https://registry.npmjs.org/",
-    "access": "public",
-    "types": "./dist/index.d.ts"
-  }
-}
