--- packages/builder/plugin-source-build/tests/setup.ts
@@ -1,4 +1,0 @@
-import { expect } from 'vitest';
-import { createSnapshotSerializer } from '@scripts/vitest-config';
-
-expect.addSnapshotSerializer(createSnapshotSerializer());
