--- packages/builder/plugin-source-build/tests/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "@modern-js/tsconfig/base",
-  "compilerOptions": {
-    "noEmit": true,
-    "declaration": false,
-    "jsx": "preserve",
-    "baseUrl": "./",
-    "isolatedModules": true,
-    "paths": {}
-  },
-}
