--- packages/builder/plugin-source-build/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "extends": "@modern-js/tsconfig/base",
-  "compilerOptions": {
-    "target": "ES2019",
-    "declaration": true,
-    "outDir": "./dist",
-    "jsx": "preserve",
-    "baseUrl": "./",
-    "isolatedModules": true,
-    "paths": {},
-    "skipLibCheck": true
-  },
-  "include": ["src"]
-}
